requirements.txt
kafka-python
pyspark==3.1.2
elasticsearch[async]
faker
pandas
prefect